requirements.txt
Pillow
requests
requests_toolbelt
